Add block preconditioners
In this commit, block preconditioners for the coupled, constant
coefficients linear Darcy system are added to the module.

All three (BlockDiagonal, BlockTriangular, and
BlockTriangularFactorization) strategies are implemented.

For the velocity block only the InverseMassMatrix preconditioner,
and for the pressure block only the LaplaceOperator are supported.
The Laplace system is subject to a single multigrid V-cycle, exact
inversion is not yet supported.