Implement spatially variable coefficients
With this commit, spatially variable permeability and porosity
fields are now possible in terms of analytical field functions.

To realize a field for the inverse permeability field, which is a
second-rank tensor, function evaluator methods for symmetric and
non-symmetric second-rank tensors are added that work with
vectorized arrays.

Preconditioners have not been adapted to the spatially coefficients
yet.