Implement ALE for momentum balance
This commit implements the ALE and coupling terms for the momentum balance equations.
As of now, only a function mesh movement type is accepted.
An ale momentum operator kernel is added which manages the grid velocity. It is embedded
into the momentum operator.  This kernel is also used to implement the RHS contribution of the
permeability term with grid velocity. 

Plus some small adjustments and beauty.