Implement porosity coupling for divergence operator
This commit implements the relation between porosity and deformation for the divergence operator.
For every Gauss point, the porosity is calculated beforehand at every Gauss point (i.e. not vectorized) at every evaluation step
with a corresponding FE(Face)Values.
This is expected to cause significant performance loss, which is accepted for now.